+++ test/tag_field_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:tag_field/tag_field.dart';
-import 'package:tag_field/tag_provider.dart';
 
 void main() {
   test('TagInputNotifier adds and removes tags correctly', () {
@@ -8,7 +7,7 @@ void main() {
     final notifier = TagInputNotifier(widget);
 
     // Test adding a tag
-    // notifier.addTag(false, 'test',BuildContext context); // No context needed
+    notifier.addTag(false, 'test', null);
     expect(notifier.tags, ['test']);
 
     // Test removing a tag